0.3.0
Transaformation of js files in module : extention = mjs. Add a square class which herit from polygon class. A square can be defnied by its vertices, or by a center and a direction which can be more commod. This class will be usefull to model and manipulate some hitboxes in games